printf format string fixups.

+++ src/tools/owmap/bspfile_abstract.c
@@ -450,7 +450,7 @@ void PrintBSPFileSizes(void)
 	Sys_Printf("\n");
 
 	Sys_Printf("%9d lightmaps     %9d\n", numBSPLightBytes / (game->lightmapSize * game->lightmapSize * 3), numBSPLightBytes);
-	Sys_Printf("%9d lightgrid     %9d *\n", numBSPGridPoints, (int)(numBSPGridPoints * sizeof(*bspGridPoints)));
+	Sys_Printf("%9d lightgrid     %9d *\n", (int)numBSPGridPoints, (int)(numBSPGridPoints * sizeof(*bspGridPoints)));
 	Sys_Printf("          visibility    %9d\n", numBSPVisBytes);
 }
